docs: add second example of use-count-down hook
Also fixed some grammar and little details
